requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==11.0.3
openai>=1.12.0
python-multipart==0.0.6
python-dotenv==1.0.0
SpeechRecognition==3.10.0
pydub==0.25.1
